Seção 9-Dto e Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,11 +1,10 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using Microsoft.AspNetCore.Mvc;
-using ProEventos.Domain;
 using ProEventos.Application.Contratos;
 using System.Threading.Tasks;
 using System;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -30,7 +29,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos ==null) return NotFound("Nenhum evento encontrado. ");
+                if (eventos ==null) return NoContent();
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -45,7 +44,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetEventoByIdAsync(id,true);
-                if (eventos ==null) return NotFound("Nenhum evento encontrado. ");
+                if (eventos ==null) return NoContent();
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -60,7 +59,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosByTemaAsync(tema,true);
-                if (eventos ==null) return NotFound("Eventos por tema não encontrados.");
+                if (eventos ==null) return NoContent();
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -70,12 +69,12 @@ namespace ProEventos.API.Controllers
             
         }
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var eventos = await _eventoService.AddEventos(model);
-                if (eventos ==null) return BadRequest("Erro ao tentar adicionar evento.");
+                if (eventos ==null) return NoContent();
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -84,12 +83,12 @@ namespace ProEventos.API.Controllers
             }
         }
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var eventos = await _eventoService.UpdateEvento(id, model);
-                if (eventos ==null) return BadRequest("Erro ao tentar editar evento.");
+                if (eventos ==null) return NoContent();
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -102,9 +101,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento ==null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
                 Ok("Deletado") : 
-                BadRequest("Evento não deletado");
+                throw new Exception("Ocorre um problema não específico ao tentar delatar o evento");
                 
             }
             catch (Exception ex)
